Reorganized Signer and Participant unit tests
The tests covering the logic attached to the Participant struct have
been moved to participant_test.go and operate on the Participant
structure. The commitment validation logic in the
validateGroupCommitmentsBase has been covered with unit tests and tests
for validateGroupCommitments have been simplified to not repeat already
performed tests.
